capstone_backend-main/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Authentication & Security
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0

# Form data
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2

# AI/ML - Optional
google-generativeai==0.3.2
